app/requirements.txt
# -- Web Framework -------------------------------------------------------------
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
python-multipart>=0.0.9

# -- Vector Database -----------------------------------------------------------
qdrant-client>=1.9.0

# -- Embeddings (BGE-m3 via plain HuggingFace transformers + torch) -----------
# TRANSFORMERS_NO_TF=1 env var prevents any TensorFlow/Keras imports.
transformers>=4.36.0
torch>=2.2.0

# -- Data Processing -----------------------------------------------------------
numpy>=1.26.0
pandas>=2.2.0

# -- Database ------------------------------------------------------------------
psycopg2-binary>=2.9.9
sqlalchemy>=2.0.0

# -- AWS (Athena + Iceberg ingestion, S3 access) --------------------------------
boto3>=1.34.0
pyathena>=3.3.0

# -- Validation & Config -------------------------------------------------------
pydantic>=2.7.0
pydantic-settings>=2.2.0
python-dotenv>=1.0.0

# -- Caching -------------------------------------------------------------------
cachetools>=5.3.0

# -- Progress Bar (for ingestion / training scripts) ---------------------------
tqdm>=4.66.0
